Add sanitization protocol for structs
Sometimes, we don't want to store sensitive data in the tracing
software. We also want to ensure that data can be safely
stored in usually JSON-like storage. At the same time, conversion to
JSON is a contextual - conversion for internal API is different than
the conversion for logs.

That means that we need a protocol implemented _before_ converting
data to JSON to sanitize the data.

Signed-off-by: Maciej Szlosarczyk <maciej@mindvalley.com>