Do not record field resolution events from GraphQL
The explanation for this is a little bit complex. The collector
written in go provides interesting functionality to sample events
based on various criteria (status, probabilistic distribution etc)
but it operates on the level of trace (a graph of many spans) and it
can drop the whole graph all at once.

For our noisy GraphQL events, we actually want the trace to be
preserved, but at the same time, do not record too many spans when it
is not necessary.

This introduces a flag that does exactly that - by default, all spans
from field resolution events will be dropped and not taken into
account. We can enable them by passing `include_field_resolution:
true` to the absinthe tracer initialization.

Signed-off-by: Maciej Szlosarczyk <maciej@mindvalley.com>