Add context propagation for Ecto preloads
When Ecto runs more than one preload, it runs it in a Task. To properly
attach the that span into its parent, we need to look at ancestors of
that process. There is now an upstream library that does exactly that,
so no longer will ecto preloads be orphaned spans.

Signed-off-by: Maciej Szlosarczyk <maciej@mindvalley.com>